Further cleanup work

+++ PerFi.API/Requests/CreateFinanceSnapshotRequest.cs
@@ -2,4 +2,4 @@ namespace PerFi.API.Requests;
 
 public record CreateFinanceSnapshotRequest(
     DateOnly SnapshotDate,
-    IReadOnlyDictionary<int, double> AccountIdToBalanceMap);
+    IReadOnlyDictionary<int, decimal> AccountIdToBalanceMap);
