+++ PerFi.API/Responses/AccountBalanceResponse.cs
@@ -2,4 +2,4 @@ namespace PerFi.API.Responses;
 
 public sealed record AccountBalanceResponse(
     AccountResponse Account,
-    double Balance);
+    decimal Balance);
