+++ PerFi.Application/Commands/CreateFinanceSnapshotCommand.cs
@@ -2,4 +2,4 @@ namespace PerFi.Application.Commands;
 
 public record CreateFinanceSnapshotCommand(
     DateOnly SnapshotDate,
-    IReadOnlyDictionary<int, double> AccountIdToBalanceMap);
+    IReadOnlyDictionary<int, decimal> AccountIdToBalanceMap);
